Turn off verbose screen when multiprocessing

+++ pyFAT_astro/main.py
@@ -209,7 +209,7 @@ def main(argv):
             #screen out put becomes a mess in multiprocessing so we turn it of
             Original_Configuration['VERBOSE_SCREEN'] = False
             ids,no_processes,no_galaxies = sf.calculate_number_processes(Original_Configuration)
-            to_maps = [(x,Original_Configuration,Full_Catalogue,100.*(x-Original_Configuration['CATALOGUE_START_ID'])/no_galaxies) for x in ids]
+            to_maps = [(x,Original_Configuration,Full_Catalogue,float(100.*(x-Original_Configuration['CATALOGUE_START_ID'])/no_galaxies)) for x in ids]
             #not issueing all galaxies individually in the pool leads to some splits taking a long time for which we then have to wait while others do nothing
             with get_context("spawn").Pool(processes=no_processes) as pool:
                 results = pool.starmap(FAT_Galaxy_Loops, to_maps)
